# src/openapi.json
{"openapi":"3.1.0","info":{"title":"CSCE 548 Project 2 Services","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/anime":{"get":{"summary":"Api List Anime","operationId":"api_list_anime_anime_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Api Create Anime","operationId":"api_create_anime_anime_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/anime/{anime_id}":{"get":{"summary":"Api Get Anime","operationId":"api_get_anime_anime__anime_id__get","parameters":[{"name":"anime_id","in":"path","required":true,"schema":{"type":"integer","title":"Anime Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Api Update Anime","operationId":"api_update_anime_anime__anime_id__put","parameters":[{"name":"anime_id","in":"path","required":true,"schema":{"type":"integer","title":"Anime Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","title":"Payload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Api Delete Anime","operationId":"api_delete_anime_anime__anime_id__delete","parameters":[{"name":"anime_id","in":"path","required":true,"schema":{"type":"integer","title":"Anime Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/studios":{"get":{"summary":"Api List Studios","operationId":"api_list_studios_studios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Api Create Studio","operationId":"api_create_studio_studios_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/studios/{studio_id}/country":{"put":{"summary":"Api Update Studio Country","operationId":"api_update_studio_country_studios__studio_id__country_put","parameters":[{"name":"studio_id","in":"path","required":true,"schema":{"type":"integer","title":"Studio Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","title":"Payload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/studios/{studio_id}":{"delete":{"summary":"Api Delete Studio","operationId":"api_delete_studio_studios__studio_id__delete","parameters":[{"name":"studio_id","in":"path","required":true,"schema":{"type":"integer","title":"Studio Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/genres":{"get":{"summary":"Api List Genres","operationId":"api_list_genres_genres_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Api Create Genre","operationId":"api_create_genre_genres_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/anime/{anime_id}/genres/{genre_id}":{"post":{"summary":"Api Add Genre","operationId":"api_add_genre_anime__anime_id__genres__genre_id__post","parameters":[{"name":"anime_id","in":"path","required":true,"schema":{"type":"integer","title":"Anime Id"}},{"name":"genre_id","in":"path","required":true,"schema":{"type":"integer","title":"Genre Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Api Remove Genre","operationId":"api_remove_genre_anime__anime_id__genres__genre_id__delete","parameters":[{"name":"anime_id","in":"path","required":true,"schema":{"type":"integer","title":"Anime Id"}},{"name":"genre_id","in":"path","required":true,"schema":{"type":"integer","title":"Genre Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}